requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.0
typer==0.9.0
rich==13.7.0

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.10.0
pymupdf==1.23.14

# AI and NLP
openai==1.6.1
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.2.2
spacy==3.7.2
nltk==3.8.1

# Data processing
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2

# Report generation
reportlab==4.0.8
python-docx==1.1.0
jinja2==3.1.2
matplotlib==3.8.2
seaborn==0.13.0

# Web interface (optional)
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.29.0

# Utilities
tqdm==4.66.1
pathlib2==2.3.7
python-magic==0.4.27
